ci: tasklist extracted profile mock to visual test setup
