+++ tasklist/client/e2e/visual-fixtures.ts
@@ -9,6 +9,17 @@
 import {test as base} from './test-fixtures';
 import {type Page} from '@playwright/test';
 
+const MOCK_TENANTS = [
+  {
+    id: 'tenantA',
+    name: 'Tenant A',
+  },
+  {
+    id: 'tenantB',
+    name: 'Tenant B',
+  },
+];
+
 type PlaywrightFixtures = {
   page: Page;
 };
@@ -27,7 +38,23 @@ const test = base.extend<PlaywrightFixtures>({
         }),
       }),
     );
-
+    await page.route('**/v2/authentication/me', (route) =>
+        route.fulfill({
+          status: 200,
+          body: JSON.stringify({
+            userId: 'demo',
+            displayName: 'demo',
+            permissions: ['READ', 'WRITE'],
+            salesPlanType: null,
+            roles: null,
+            c8Links: [],
+            tenants: MOCK_TENANTS,
+          }),
+          headers: {
+            'content-type': 'application/json',
+          },
+        }),
+    );
     await page.route('**/client-config.js', (route) =>
       route.fulfill({
         status: 200,
@@ -57,4 +84,4 @@ const test = base.extend<PlaywrightFixtures>({
   },
 });
 
-export {test};
+export {test, MOCK_TENANTS};
